+++ src/Driver/TcpdiDriver.php
@@ -2,10 +2,10 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf\Driver;
+namespace kadudutra\libmergepdf\Driver;
 
-use iio\libmergepdf\Exception;
-use iio\libmergepdf\Source\SourceInterface;
+use kadudutra\libmergepdf\Exception;
+use kadudutra\libmergepdf\Source\SourceInterface;
 
 final class TcpdiDriver implements DriverInterface
 {