+++ src/Merger.php
@@ -2,13 +2,13 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf;
+namespace kadudutra\libmergepdf;
 
-use iio\libmergepdf\Driver\DriverInterface;
-use iio\libmergepdf\Driver\DefaultDriver;
-use iio\libmergepdf\Source\SourceInterface;
-use iio\libmergepdf\Source\FileSource;
-use iio\libmergepdf\Source\RawSource;
+use kadudutra\libmergepdf\Driver\DriverInterface;
+use kadudutra\libmergepdf\Driver\DefaultDriver;
+use kadudutra\libmergepdf\Source\SourceInterface;
+use kadudutra\libmergepdf\Source\FileSource;
+use kadudutra\libmergepdf\Source\RawSource;
 
 /**
  * Merge existing pdfs into one