+++ src/Pages.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf;
+namespace kadudutra\libmergepdf;
 
 /**
  * Parse page numbers from string