+++ src/Source/FileSource.php
@@ -2,11 +2,11 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf\Source;
+namespace kadudutra\libmergepdf\Source;
 
-use iio\libmergepdf\PagesInterface;
-use iio\libmergepdf\Pages;
-use iio\libmergepdf\Exception;
+use kadudutra\libmergepdf\PagesInterface;
+use kadudutra\libmergepdf\Pages;
+use kadudutra\libmergepdf\Exception;
 
 /**
  * Pdf source from file