+++ src/Source/SourceInterface.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf\Source;
+namespace kadudutra\libmergepdf\Source;
 
-use iio\libmergepdf\PagesInterface;
+use kadudutra\libmergepdf\PagesInterface;
 
 interface SourceInterface
 {