+++ tests/Driver/TcpdiDriverTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types = 1);
 
-namespace iio\libmergepdf\Driver;
+namespace kadudutra\libmergepdf\Driver;
 
-use iio\libmergepdf\Source\SourceInterface;
-use iio\libmergepdf\Exception;
-use iio\libmergepdf\Pages;
+use kadudutra\libmergepdf\Source\SourceInterface;
+use kadudutra\libmergepdf\Exception;
+use kadudutra\libmergepdf\Pages;
 
 class TcpdiDriverTest extends \PHPUnit\Framework\TestCase
 {