+++ CHANGELOG.md
@@ -104,7 +104,7 @@ After some early deviations this project now adheres to [Semantic Versioning](ht
 ## [2.1] - 2013-01-31
 
 ### Changed
-- Moved to `iio` namespace.
+- Moved to `kadudutra` namespace.
 
 ## [2.0] - 2012-10-02